Refactoring: simplify testing
